Update to have FS not ready by default, and make changes suggested by code review.  Moving setCookieNamesAsUserVars to script,
