+++ src/set-cookies.js
@@ -1,5 +1,26 @@
-import {registerFsReady} from "./utils/fs";
-import {setCookiesAsUserVars} from "./utils/cookie";
+import {fs, registerFsReady} from "./utils/fs";
+import {getAvailableCookies} from "./utils/cookie";
+
+/**
+ * Set fs userVars based on cookie names.  Will be set as cookie.name=cookie.value
+ * If cookie is not available, it will not be in the userVars, and a warning will be printed
+ * Also note that userVars in FullStory cannot start with _ character, so make sure your cookies
+ * don't start with that otherwise they won't appear.
+ * @param cookieNames An array of cookie names to set as values
+ */
+function setCookiesAsUserVars( cookieNames ){
+  let cookies = getAvailableCookies();
+  let userVars = {};
+  for( let i=0; i<cookieNames.length; i++ ){
+    if( cookieNames[i] in cookies ){
+      userVars[cookieNames[i]] = cookies[cookieNames[i]];
+    }
+    else {
+      fs( "log" )( "warn", "Cookie " + cookieNames[i] + " was not found when trying to setCookiesAsUserVars call" );
+    }
+  }
+  fs( "setUserVars" )( userVars );
+}
 
 /**
  *
